Revised "key advantages" for Mule Gateway
I find the embedded nature of the Mule Gateway provides significant security advantages in cases where applications might be deployed into different topologies over time. It's worth pointing this out in the docs, I believe.

Note that the reference to Anypoint API Governance is not hyperlinked. I don't know the markup for that, but suggest that link be made.